perf(learn): lazy-load the article renderer + corpus off the EN fast path
What
- src/learn-hub/content-model.js becomes a thin barrel over two new modules:
  content-structure.js (section ids + translation keys, eager-safe) and
  content-text.js (the localized EN+AR copy corpus). The barrel preserves the
  exact export surface, so the Node static-fallback generator
  (scripts/node/render-learn-static-fallback.mjs) and every existing consumer
  keep working unchanged.
- src/pages/learn.js drops its static article-renderer import; the renderer
  (which is the only importer of content-text) now loads via a cached dynamic
  import behind mountArticleExperience. The EN static-fallback fast path never
  triggers it. A monotonic mount token prevents a slow lazy render from
  clobbering a newer language toggle; load failure falls back to the honest
  missing-article state.
- toc-renderer/content-registry point at content-structure so the EN toc
  enhancement keeps its synchronous structural data.
- New tests/learn-corpus-split.test.js (import-graph guards on real import
  statements, barrel contract, structural-key -> en+ar copy resolution) and
  tests/e2e/learn-corpus-lazy.spec.js (EN fetches no corpus chunk; AR renders
  the Arabic article).

Why
Verified Work Queue item 19 phase-0 (audit F): learn.html shipped its full
bilingual article corpus + renderer eagerly (~33.3 KB gz) even on the EN path
that renders prerendered static HTML and never uses them.

Proof
- Build: eager learn chunk 33.3 KB gz -> 23.5 KB gz; renderer+corpus now an
  11.1 KB gz lazy chunk fetched only when rendering (AR path).
- Browser (built dist, chromium): EN learn.html fetches zero corpus chunks,
  full article visible from static HTML, no page errors; ?lang=ar fetches
  article-renderer chunk once and renders Arabic sections (h2s verified).
- node scripts/node/render-learn-static-fallback.mjs: exit 0, learn.html
  byte-identical (no diff).
- Gates: lint 0, quality 0, validate 0, tests 1662/1662, build 0; e2e spec
  4/4 with --repeat-each=2.

Risks
- The AR article render is now one dynamic import later than before; the
  chunk is small (11.1 KB gz) and cached. Mount-token guard covers toggle
  races; EN behavior is byte-identical (same static HTML, fewer bytes).
